+++ src/main/java/com/evacipated/cardcrawl/modthespire/patcher/ParamInfo.java
@@ -1,5 +1,6 @@
 package com.evacipated.cardcrawl.modthespire.patcher;
 
+import com.evacipated.cardcrawl.modthespire.lib.ByRef;
 import javassist.CtBehavior;
 import javassist.CtClass;
 import javassist.Modifier;
@@ -92,8 +93,16 @@ class ParamInfo
         return ctBehavior.getParameterTypes()[position - 1];
     }
 
-    String getTypename() throws NotFoundException
+    String getTypename() throws NotFoundException, ClassNotFoundException
     {
+        if (position > 0) {
+            for (Object o : ctBehavior.getParameterAnnotations()[position - 1]) {
+                if (o instanceof ByRef) {
+                    return ((ByRef) o).type();
+                }
+            }
+        }
+
         CtClass type = getType();
         if (type == null) {
             return "";