+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog ##
 #### dev ####
 * Save workshop item locations so they can be used later without Steam running
+* Fix regression of ByRef named type parameter
 
 #### v3.12.0 ####
 * Make ByRef work on all Insert patch parameters, not just localvars